Lab1/requirements.txt
Flask
chromadb
nltk
numpy
scikit-learn
PyPDF2
python-docx
gensim
langdetect
